Refactor PendingRequestStore
